% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_asl_loglik.R
\name{compute_asl_loglik}
\alias{compute_asl_loglik}
\title{Compute log likelihood in associative and structual learning model}
\usage{
compute_asl_loglik(alpha, data, n_cs, opts_asl)
}
\arguments{
\item{alpha}{concentration parameter}

\item{data}{single-subject data}

\item{n_cs}{number of CS}

\item{opts_asl}{options structure}
}
\value{
lik: log-likelihood

latents: structure containing latent variables:\preformatted{          b: beta coefficient mapping model CR to  measured CR

          sd: maximum likelihood standard deviation

          CR: predicted CR
}
}
\description{
\code{compute_asl_loglik} compute log likelihood of data under the latent cause model.
}
\examples{
# results <- compute_asl_loglik(alpha,data,n_cs,opts_asl)
}
