% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_asl_loglik.R, R/compute_lcm_loglik.R
\name{compute_lcm_loglik}
\alias{compute_lcm_loglik}
\title{Compute log likelihood}
\usage{
compute_lcm_loglik(alpha, data, n_cs, opts)

compute_lcm_loglik(alpha, data, n_cs, opts)
}
\arguments{
\item{alpha}{concentration parameter}

\item{data}{single-subject data}

\item{n_cs}{number of CS}

\item{opts}{options structure}

\item{alpha}{concentration parameter}

\item{data}{single-subject data}

\item{n_cs}{number of CS}

\item{opts}{options structure}
}
\value{
lik: log-likelihood

latents: structure containing latent variables:\preformatted{          b: beta coefficient mapping model CR to  measured CR

          sd: maximum likelihood standard deviation

          CR: predicted CR
}

lik: log-likelihood

latents: structure containing latent variables:\preformatted{          b: beta coefficient mapping model CR to  measured CR

          sd: maximum likelihood standard deviation

          CR: predicted CR
}
}
\description{
\code{compute_lcm_loglik} compute log likelihood of data under the latent cause model.

\code{compute_lcm_loglik} compute log likelihood of data under the latent cause model.
}
\examples{
# results <- compute_lcm_loglik(alpha,data,n_cs,opts)
# results <- compute_lcm_loglik(alpha,data,n_cs,opts)
}
