% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_negative_loglike.R
\name{compute_loglike}
\alias{compute_loglike}
\title{Compute log likelihood of model selected}
\usage{
compute_loglike(param, data, model, opts)
}
\arguments{
\item{param}{parameter}

\item{data}{single-subject data}

\item{model}{1 = latent cause model, 2 = latent cause modulated RW model}

\item{opts}{options}
}
\value{
lik: log-likelihood

latents: structure containing latent variables:\preformatted{          b: beta coefficient mapping model CR to  measured CR

          sd: maximum likelihood standard deviation

          CR: predicted CR
}
}
\description{
\code{compute_loglike} compute log likelihood of data under the latent cause model.
}
