% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_negative_loglike.R
\name{compute_negative_loglike}
\alias{compute_negative_loglike}
\title{Compute negative log likelihood of model selected}
\usage{
compute_negative_loglike(param, data, model, opts)
}
\arguments{
\item{param}{parameter}

\item{data}{single-subject data}

\item{model}{1 = latent cause model, 2 = latent cause modulated RW model}

\item{opts}{options}
}
\value{
negative log likelihood
}
\description{
\code{compute_negative_loglike} compute negative log likelihood of data under the latent cause model.
}
