\name{Chlamydia}
\docType{data}
\alias{Chlamydia}
\title{Data sets}
\description{
Results on 4 diagnostic tests  for Chlamydia trachomatis.
The complete data, called Chlamydia, has results for 3551 patients.
}
\usage{
Chlamydia
sample.Chlamydia
}

\format{ 
Both data frames contain the following columns:
\itemize{
\item{LCR: }{ Ligase Chain Reaction test (LCR) (0: negative; 1: positive)}
\item{PCR: }{ Polymerase Chain Reaction test (PCR) (0: negative; 1: positive)}
\item{DNAP: }{ DNA probe test (DNAP) (0: negative; 1: positive)}
\item{CUL: }{ tissue culture test(CUL) (0: negative; 1: positive)}
\item{count: }{ number of individuals with the test results of X1, X2, X3 and X4.}
}
}

\keyword{datasets}

