% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcm.R
\name{estimate_by_optim}
\alias{estimate_by_optim}
\title{Estimate parameters of LCM by optim}
\usage{
estimate_by_optim(data, model, opts, parameter_range)
}
\arguments{
\item{data}{data from fit_asl or fit_asl_parallel}

\item{model}{1 = latent cause model, 2 = latent cause modulated RW model}

\item{opts}{(optional) structure defining ASL options}

\item{parameter_range}{(optional) range of parameter}
}
\value{
estimated parameters, negative log likelihood
}
\description{
\code{estimate_by_optim} is function to estimate lcm parameter using optim
}
