% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcm.R
\name{estimate_by_post_mean}
\alias{estimate_by_post_mean}
\title{Estimate alpha of LCM by post mean}
\usage{
estimate_by_post_mean(data, model, opts, alpha)
}
\arguments{
\item{data}{data from fit_asl or fit_asl_parallel}

\item{model}{1 = latent cause model, 2 = latent cause modulated RW model}

\item{opts}{(optional) structure defining ASL options}

\item{alpha}{vector of alpha}
}
\value{
df data frame containing post_mean_alpha(posterior mean alpha),
logBF(log Bayes factor for the alpha>=0 model relative to the alpha=0 model),
loglikelihood and probability each alpha.
}
\description{
\code{estimate_by_post_mean} is function to estimate alpha and logBF
}
