% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lcm_a.R
\name{estimate_lcm_a}
\alias{estimate_lcm_a}
\title{Estimate alpha and logBF}
\usage{
estimate_lcm_a(data, n_cs, opts, alpha)
}
\arguments{
\item{data}{data from LCM_pfit}

\item{n_cs}{number of CS}

\item{opts}{(optional) structure defining LCM options (see LCM_opts)}

\item{alpha}{vector of alpha}
}
\value{
df data frame containing post_mean_alpha(posterior mean alpha),
logBF(log Bayes factor for the alpha>=0 model relative to the alpha=0 model),
loglikelihood and probability each alpha.
}
\description{
\code{estimate_lcm_a} is function to estimate alpha and logBF
}
