% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_asl.R
\name{fit_asl}
\alias{fit_asl}
\title{Fit data to associative and structural learning model}
\usage{
fit_asl(data, n_cs, asl_opts, set_alpha)
}
\arguments{
\item{data}{long format data containing the following variables
(Order and name is exactly the same as following):\preformatted{   ID  Subject ID

   Time stimulus onset time

   CR  Conditioned Response

   US  Unconditioned Stimulus

   CS  Conditioned Stimului. If using multiple CS, set variables name as CS1,CS2,CS3...
}}

\item{n_cs}{number of CS}

\item{asl_opts}{(optional)}

\item{set_alpha}{(optional) option of alpha values to evaluate.
You can change max value and numbers of alpha as followings(default: max value = 10, number = 50).
list(num_alpha = 100, max_alpha = 5)}
}
\value{
return the fit, parameters and plc_vcr.
fit is fitting results. parameters is parameters estmated including post_mean_alpha(posterior mean alpha) and
logBFlog(Bayes factor for the alpha>=0 model relative to the alpha=0 model).
plc_vcr is matrix of latent cause posterior and V & CR predicted.
}
\description{
\code{fit_asl} fit latent cause model to conditioning data
}
\examples{

# results <- fit_asl(data,n_cs,asl_opts,list(num_alpha = 100, max_alpha = 10))
}
