% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcm_parallel.R
\name{fit_lcm_parallel}
\alias{fit_lcm_parallel}
\title{Fit data to latent cause model using parallel computing}
\usage{
fit_lcm_parallel(data, n_cs, opts, set_alpha)
}
\arguments{
\item{data}{long format data containing the following variables
(Order and name is exactly the same as following):\preformatted{   ID  Subject ID

   CR  Conditioned Response

   US  Unconditioned Stimulus

   CS  Conditioned Stimului. If you use a multiple CS, set variables name as CS1,CS2,CS3...
}}

\item{n_cs}{number of CS}

\item{opts}{(optional) LCM options (see LCM_opts)}

\item{set_alpha}{(optional) option of alpha values to evaluate.
You can change max value and numbers of alpha as followings(default: max value = 10, number = 50).
list(num_alpha = 100, max_alpha = 5)}
}
\value{
data post_mean_alpha(posterior mean alpha) and
logBFlog(Bayes factor for the alpha>=0 model relative to the alpha=0 model)
add to original data
}
\description{
\code{fit_lcm_parallel} fit latent cause model to conditioning data using parallel computing
}
\examples{

# results <- fit_lcm_parallel(data,n_cs,opts,list(num_alpha = 100, max_alpha = 10))
}
