% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_lcm.R
\name{infer_lcm}
\alias{infer_lcm}
\title{Latent cause model using particle filtering or local maximum a posteriori inference}
\usage{
infer_lcm(X, opts)
}
\arguments{
\item{X}{matrix of stimulus inputs consisting of the number of trial rows and
the number of stimulus features columns. The first feature (column 1) is the US,
and the rest of the features (column 2 through D) are CSs.}

\item{opts}{(optional)structure containing various options (see LCM_opts).
Missing fields are set to defaults. If opts.M = 1, then the model computes a local MAP estimate.}
}
\value{
opts: options used in inference

V: US prediction each trial

post: matrix of latent cause posterior consisting of the number of trial rows and
the probability of latent cause k being active on trial t, after observing the all the features.
K (the number of latent causes) is determined adaptively by the model.
}
\description{
\code{infer_lcm} conduct Particle filtering or local maximum a posteriori inference
for latent cause model of associative learning
}
\examples{
# results <- infer_lcm(X,opts)
}
