% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_alpha.r
\name{LCM_alpha}
\alias{LCM_alpha}
\title{Function to estimate alpha and logBF}
\usage{
LCM_alpha(data, n_cs, opts, alpha)
}
\arguments{
\item{data}{data from LCM_pfit}

\item{n_cs}{number of CS}

\item{opts}{(optional) structure defining LCM options (see LCM_opts)}

\item{alpha}{vector of alpha}
}
\value{
df data frame containing post_mean_alpha(posterior mean alpha) and
logBF(: )log Bayes factor for the alpha>=0 model relative to the alpha=0 model)
}
\description{
\code{LCM_alpha} is function to estimate alpha and logBF
}
