% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_fit.R
\name{LCM_fit}
\alias{LCM_fit}
\title{Fit latent cause model to data.
INPUTS:
data - long format data containing the following variables:
ID: Subject ID
CR: Conditioned Response
US: Unconditioned Stimulus
CS: Conditioned Stimului(CS1,CS2,CS3...)
n_cs - number of CS
opts (optional) - structure defining LCM options (see LCM_opts.m)
OUTPUTS:
results - \link{nSubjects x 1} structure containing the following fields:
.alpha - concentration parameter values
.P - posterior probability distribution over alpha
.lik - log-likelihood for each alpha value
.latents - latent variables for each alpha value (see LCM_infer)
.logBF - log Bayes factor for the alpha>=0 model relative to the alpha=0 model}
\usage{
LCM_fit(data = data, n_cs = n_cs, opts = opts)
}
\description{
Fit latent cause model to data.
INPUTS:
data - long format data containing the following variables:
ID: Subject ID
CR: Conditioned Response
US: Unconditioned Stimulus
CS: Conditioned Stimului(CS1,CS2,CS3...)
n_cs - number of CS
opts (optional) - structure defining LCM options (see LCM_opts.m)
OUTPUTS:
results - \link{nSubjects x 1} structure containing the following fields:
.alpha - concentration parameter values
.P - posterior probability distribution over alpha
.lik - log-likelihood for each alpha value
.latents - latent variables for each alpha value (see LCM_infer)
.logBF - log Bayes factor for the alpha>=0 model relative to the alpha=0 model
}
\examples{
results <- LCM_fit(data,opts)
}
