% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_infer.R
\name{LCM_infer}
\alias{LCM_infer}
\title{Particle filtering or local maximum a posteriori inference for latent cause model of associative learning.}
\usage{
LCM_infer(X = X, opts = opts)
}
\description{
Particle filtering or local maximum a posteriori inference for latent cause model of associative learning.
}
\examples{
results <- LCM(X,opts)
}
