% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_lik.R
\name{LCM_lik}
\alias{LCM_lik}
\title{Compute log-likelihood of data under the latent cause model.}
\usage{
LCM_lik(alpha = alpha, data = data, n_cs = n_cs, opts = opts)
}
\description{
INPUTS:
alpha - concentration parameter
data - single-subject data
opts - options structure
OUTPUTS:
lik - log-likelihood
latents - structure containing latent variables:
.b - beta coefficient mapping model CR to  measured CR
.sd - maximum likelihood standard deviation
.CR - predicted CR
}
\examples{
LCM_lik(alpha,data,[opts])
}
