% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_pfit.R
\name{LCM_pfit}
\alias{LCM_pfit}
\title{Fit latent cause model using parallel computing}
\usage{
LCM_pfit(data, n_cs, opts, parallel = FALSE)
}
\arguments{
\item{data}{long format data containing the following variables
(Order and name is exactly the same as following):\preformatted{   ID  Subject ID

   CR  Conditioned Response

   US  Unconditioned Stimulus

   CS  Conditioned Stimului. If using multiple CS, set variables name as CS1,CS2,CS3...
}}

\item{n_cs}{number of CS}

\item{opts}{(optional) structure defining LCM options (see LCM_opts)}

\item{parallel}{If set TRUE(default is FALSE), using parallel processing}
}
\value{
data post_mean_alpha(posterior mean alpha) and
logBFlog(Bayes factor for the alpha>=0 model relative to the alpha=0 model)
add to original data
}
\description{
\code{LCM_fit} fit latent cause model to conditioning data using parallel computing
}
\examples{

# results <- LCM_pfit(data,n_cs,opts,parallel=TRUE)
}
