% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_pfit.R
\name{LCM_pfit_single}
\alias{LCM_pfit_single}
\title{Function to fit latent cause model to data for parallel computing}
\usage{
LCM_pfit_single(data, n_cs, opts, alpha)
}
\arguments{
\item{data}{long format data containing the following variables
(Order and name is exactly the same as following):\preformatted{   ID  Subject ID

   CR  Conditioned Response

   US  Unconditioned Stimulus

   CS  Conditioned Stimului. If using multiple CS, set variables name as CS1,CS2,CS3...
}}

\item{n_cs}{number of CS}

\item{opts}{(optional) structure defining LCM options (see LCM_opts)}

\item{alpha}{vector of alpha}
}
\value{
df data frame containing post_mean_alpha(posterior mean alpha) and
logBF(: )log Bayes factor for the alpha>=0 model relative to the alpha=0 model)
}
\description{
\code{LCM_fit} is function to fit latent cause model to data for parallel computing
}
