% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associative_structure_learning.R
\name{learn_associative_structure}
\alias{learn_associative_structure}
\title{Latent cause-modulated Rescorla-Wagner model}
\usage{
learn_associative_structure(X, opts_asl)
}
\arguments{
\item{X}{matrix cotaining time(stimulus onset, unit is sec), US, CS.}

\item{opts_asl}{(optional)list containing various options.
If you don't want to use  a nonlinear sigmoidal transformation, you set opts_asl$nst = 0.}
}
\value{
V: vector of conditioned response on each trial

Zp: latent cause posterior before observing US(Trial*K)

Z: latent cause posterior(Trial*K)
}
\description{
\code{learn_associative_structure} conduct local maximum a posteriori inference
for associative and structuallearning
}
\examples{
# results <- ilearn_associative_structure(X, opts_asl)

}
