\name{sample.Chlamydia}
\docType{data}
\alias{sample.Chlamydia}
\title{Data sets}
\description{ 
A \eqn{10\%}{} random sample of the complete data set, \code{\link{Chlamydia}},
is called sample.chlamydia and has results for 354 patients. 
}
\usage{
sample.Chlamydia
}

\format{ 
Both data frames contain the following columns:
\itemize{
\item{LCR: }{ Ligase Chain Reaction test (LCR) (0: negative; 1: positive)}
\item{PCR: }{ Polymerase Chain Reaction test (PCR) (0: negative; 1: positive)}
\item{DNAP: }{ DNA probe test (DNAP) (0: negative; 1: positive)}
\item{CUL: }{ tissue culture test(CUL) (0: negative; 1: positive)}
\item{count: }{ number of individuals with the test results of X1, X2, X3 and X4.}
}
}
\seealso{ \code{\link{Chlamydia}}  }
\keyword{datasets}

