% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcm.R
\name{set_opts}
\alias{set_opts}
\title{Set options}
\usage{
set_opts(model, opts)
}
\arguments{
\item{model}{1 = latent cause model, 2 = latent cause modulated RW model}

\item{opts}{options used in inference}
}
\value{
opts
}
\description{
\code{set_opts} set options
}
\examples{

# opts = set_opts(1, opts = list(M = 1000, c_alpha = 1))
}
