% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcm.R
\name{fit_lcm}
\alias{fit_lcm}
\title{Fit data to latent cause model}
\usage{
fit_lcm(data, model, opts, parameter_range, estimation_method)
}
\arguments{
\item{data}{long format data containing the following variables
(Order and name is exactly the same as following):\preformatted{   ID  Subject ID

   CR  Conditioned Response

   US  Unconditioned Stimulus

   CS  Conditioned Stimului or Context. If using multiple CS, set variables name as CS1,CS2,CS3...

   If you want to use LCM-RW, you have to add time variable(stimulus onset, unit is sec) before US.
}}

\item{model}{1 = latent cause model, 2 = latent cause modulated RW model}

\item{opts}{(optional) structure defining ASL options}

\item{parameter_range}{(optional)  range of parameter(a_L, a_U, e_L, e_U)}

\item{estimation_method}{(optional)  0 = optim or optimize(lcm), 1 = post mean(only latent cause model)}
}
\value{
return the fit, parameters and plc_vcr.
fit is fitting results. parameters is parameters estmated including post_mean_alpha(posterior mean alpha) and
logBFlog(Bayes factor for the alpha>=0 model relative to the alpha=0 model).
plc_vcr is matrix of latent cause posterior and V & CR predicted.
}
\description{
\code{fit_lcm} fit latent cause model to conditioning data
}
\examples{

# results <- fit_lcm(data, model, opts, parameter_range, estimation_method)
}
