% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_lcm.R
\name{infer_lcm_rw}
\alias{infer_lcm_rw}
\title{Latent cause modulated Rescorla-Wagner model}
\usage{
infer_lcm_rw(X, opts)
}
\arguments{
\item{X}{matrix cotaining time(stimulus onset, unit is sec), US, CS.}

\item{opts}{(optional)list containing various options.
If you don't want to use  a nonlinear sigmoidal transformation, you set opts$nst = 0.}
}
\value{
V: vector of conditioned response on each trial

Zp: latent cause posterior before observing US(Trial*K)

Z: latent cause posterior(Trial*K)
}
\description{
\code{infer_lcm_rw} conduct local maximum a posteriori inference
for associative and structuallearning
}
\examples{
# results <- infer_lcm_rw(X, opts)

}
