% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_fit.R
\name{LCM_fit}
\alias{LCM_fit}
\title{Fit latent cause model}
\usage{
LCM_fit(data, n_cs, opts)
}
\arguments{
\item{data}{long format data containing the following variables
(Order and name is exactly the same as following):\preformatted{   ID  Subject ID

   CR  Conditioned Response

   US  Unconditioned Stimulus

   CS  Conditioned Stimului. If using multiple CS, set variables name as CS1,CS2,CS3...
}}

\item{n_cs}{number of CS}

\item{opts}{(optional) structure defining LCM options (see LCM_opts)}
}
\value{
post_mean_alpha: posterior mean alpha

logBF: log Bayes factor for the alpha>=0 model relative to the alpha=0 model
}
\description{
\code{LCM_fit} fit latent cause model to conditioning data.
}
\examples{
# results <- LCM_fit(data,opts)
}
