% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCM_lik.R
\name{LCM_lik}
\alias{LCM_lik}
\title{Compute log-likelihood of data under the latent cause model.}
\usage{
LCM_lik(alpha = alpha, data = data, n_cs = n_cs, opts = opts)
}
\arguments{
\item{alpha}{concentration parameter}

\item{data}{single-subject data}

\item{opts}{options structure}
}
\value{
lik - log-likelihood
latents - structure containing latent variables:
.b - beta coefficient mapping model CR to  measured CR
.sd - maximum likelihood standard deviation
.CR - predicted CR
}
\description{
\code{LCM_lik}
}
\examples{
LCM_lik(alpha,data,[opts])
}
